move CLIPImageProjection to stable_diffusion
