+++ src/diffusers/__init__.py
@@ -39,7 +39,6 @@ else:
         AsymmetricAutoencoderKL,
         AutoencoderKL,
         AutoencoderTiny,
-        CLIPImageProjection,
         ControlNetModel,
         ModelMixin,
         MultiAdapter,
@@ -67,6 +66,7 @@ else:
         AutoPipelineForImage2Image,
         AutoPipelineForInpainting,
         AutoPipelineForText2Image,
+        CLIPImageProjection,
         ConsistencyModelPipeline,
         DanceDiffusionPipeline,
         DDIMPipeline,
